sane refactor and thaw frozen base model
