+++ src/ome_zarr_models/base.py
@@ -1,0 +1,7 @@
+import pydantic
+
+
+class Base(pydantic.BaseModel):
+    """
+    The base pydantic model for all metadata classes 
+    """
