+++ src/ome_zarr_models/tmp/v04/omero.py
