+++ src/ome_zarr_models/utils.py
@@ -1,8 +1,5 @@
 from typing import TypeVar
-
-
 T = TypeVar("T")
-
 
 def _unique_items_validator(values: list[T]) -> list[T]:
     for ind, value in enumerate(values, start=1):
