+++ src/ome_zarr_models/v04/models/axes.py
@@ -1,0 +1,13 @@
+from ome_zarr_models.base import Base
+
+
+class Axis(Base):
+    """
+    Model for an element of `Multiscale.axes`.
+
+    See https://ngff.openmicroscopy.org/0.4/#axes-md.
+    """
+
+    name: str
+    type: str | None = None
+    unit: str | None = None
