+++ src/ome_zarr_models/v04/omero.py
