+++ src/ome_zarr_models/v04/v04/omero.py
