--- src/ome_zarr_models/zarr_models/base.py
@@ -1,7 +1,0 @@
-import pydantic
-
-
-class FrozenBase(pydantic.BaseModel, frozen=True):
-    """
-    A frozen pydantic basemodel.
-    """
